requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.12.0
-e git+https://github.com/django-import-export/django-import-export.git@a22a278dcec327084c3b8597e20e95d38e7914ef#egg=django_import_export
django-liststyle==0.2b0
django-truncate==0.1
et-xmlfile==1.1.0
gunicorn==21.2.0
import-export==0.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
psycopg2-binary==2.9.9
PyYAML==6.0.1
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.3
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
